Больше деталей,добавлены модули
