+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "keksobooking",
+  "type":"module",
   "version": "23.0.0",
   "private": true,
   "description": "Личный проект «Кексобукинг» от HTML Academy",
